refactor to break down conversation list component
